api-pc1/requirements.txt
imagekitio
flask
pydub
sqlalchemy < 2.0
cryptography
waitress
Pillow
pymysql